afiliados-comisiones-servicio/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
aio-pika==9.4.0
